Warn instead of raising on opportunistic auth failure
If the request results in a 401, this reduces to the non-opportunistic
case, whose behavior remains the same as before.

This allows callers for whom enabling opportunistic auth is a more
appropriate default to do so, without needing to write additional
exception handling code to ensure they recover in the case that
opportunistic auth failed.  In other words, take "opportunistic" to mean
"try early, but if that fails, don't just give up right away".

Signed-off-by: Joshua Bronson <jab@twosigma.com>
[rharwood@redhat.com: restore debug message, fussy style things]