Set SanitizedResponse content to bytes instead of string
The `content` property of a `Response` object is expected to always be
bytes (the `text` property should be a string).

This change fixes an issue where consumers of the library that were
returned a `SanitizedResponse` object could fail because they expected
the `content` attribute of it to be bytes instead of a string.

Signed-off-by: Carey Metcalfe <carey@cmetcalfe.ca>